+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.1.0 / Unreleased
   * [Added] Add 'fancytree-icon' class to icon IMG
+  * [Added] #210: [ext-persist] optionally store information in localStorage
   * [Improved] [ext-columnview] css
   * [Improved]: New option 'scrollParent' is evaluated by node.scrollIntoView()
     (which is also called on expand). This allows autoScroll to work with